Conflicts:
	e-Look/src/main/java/com/e_Look/eLookEvent/control/UpdateCourseClassEvent.java

+++ e-Look/src/main/java/com/e_Look/ReportCourse/control/ReportCourseControl.java
@@ -65,7 +65,7 @@ public class ReportCourseControl extends HttpServlet {
 		}
 		
 		ReportCourseService rcService = new ReportCourseService();
-		String jsonObj = rcService.getJSON(status);
+		String jsonObj = rcService.getJSON((byte)status);
 		out.println(jsonObj);
 	}
