+++ e-Look/src/main/java/com/e_Look/ReportCourse/model/ReportCourseDAO_interface.java
@@ -10,7 +10,7 @@ public interface ReportCourseDAO_interface {
 	public ReportCourseVO findByReportId(Integer reportId);
 	public List<ReportCourseVO> getNotHandle(byte status);
 	public List<ReportCourseVO> getAll();
-	public String getJSON(Integer status);
+	public String getJSON(byte status);
 
 }
